requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
Faker==15.0.0
Markdown==3.4.1
mysqlclient==2.1.1
pycodestyle==2.9.1
Pygments==2.13.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
validate-docbr==1.10.0
